Done 29-11

+++ app/services/activity_admin_service.rb
@@ -2,7 +2,9 @@ class ActivityAdminService
     def self.getAllActivities (params)
        begin
 
-            activities = Activity.all
+            activities = Activity.left_joins(student_activity: :rating)
+            .select('activity.*, COALESCE(CEIL(AVG(rating.RatingStar)), 0) AS averageRating, COALESCE(COUNT(student_activity.ActivityCode), 0) AS numOfParticipant')
+            .group('activity.ActivityCode')
 
             # Phân trang, lọc, sắp xếp dữ liệu
             processedData = PaginationSortSearch.dataExploration(activities, params, "ActivityName")
@@ -33,17 +35,52 @@ class ActivityAdminService
         end
     end
 
-    def self.deleteActivity (activityCode)
-        begin
-            activity = Activity.find_by(ActivityCode: activityCode)
-            if activity
-                activity.destroy
-                return {success: true, message: "Xóa hoạt động thành công",status: 200}
-            else 
-                return {success: false, message: "Hoạt động không tồn tại", status: 400}
+     # Hàm kiểm tra mảng mã hoạt động đầu vào có lỗi không
+     def self.validate_input(arr_activity_code)
+        error_results = []
+        validate_result = true
+      
+        arr_activity_code.each do |activity_code|
+            activity = Activity.find_by(ActivityCode: activity_code)
+        
+            # Kiểm tra mã hoạt động có tồn tại trên hệ thống?
+            unless activity
+                error_results <<  "Không tồn tại mã hoạt động #{activity_code}" 
+                validate_result = false
+                next
             end
+        
+            # Kiểm tra hoạt động có sinh viên nào tham gia chưa
+            student_activity = StudentActivity.find_by(ActivityCode: activity_code)
+            if student_activity
+                error_results <<  "Hoạt động #{activity_code} đã có sinh viên tham gia. Không thể xóa" 
+                validate_result = false
+                next
+            end
+        
+        end
+      
+        { result: validate_result, message: error_results }
+    end 
+
+    def self.deleteActivity(arrActivityCode)
+        begin
+          validate_result = validate_input(arrActivityCode)
+      
+          if validate_result[:result]
+            arrActivityCode.each do |activityCode|
+              activity = Activity.find_by(ActivityCode: activityCode)
+              activity.destroy if activity
+            end
+            return { success: true, message: "Xóa hoạt động thành công", status: 400 }
+          else
+            return { success: false, message: validate_result[:message], status: 400 }
+          end
+        rescue ActiveRecord::InvalidForeignKey => e
+          return { success: false, message: "Không thể xóa hoạt động khi đã có sinh viên tham gia", status: 400 }
         rescue StandardError => e
-            return { success: false, message: e.message, status: 400 }
+          return { success: false, message: e.message, status: 400 }
         end
     end
+
 end
